Correct java verification classes and clarify scopes usage examples
