+++ flow-documentation/src/main/java/com/vaadin/flow/tutorial/spring/BasicSpringUsage.java
@@ -15,23 +15,8 @@
  */
 package com.vaadin.flow.tutorial.spring;
 
-import java.util.Collection;
-import java.util.Collections;
-
-import javax.servlet.ServletContext;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRegistration;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.WebApplicationInitializer;
-import org.springframework.web.context.ContextLoaderListener;
-import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
-
 import com.vaadin.flow.component.html.Div;
 import com.vaadin.flow.router.Route;
-import com.vaadin.flow.spring.SpringServlet;
-import com.vaadin.flow.spring.VaadinMVCWebAppInitializer;
 import com.vaadin.flow.tutorial.annotations.CodeFor;
 
 @CodeFor("spring/tutorial-spring-basic.asciidoc")
@@ -44,38 +29,4 @@ public class BasicSpringUsage {
         }
     }
 
-    public abstract class ExampleWebAppInitializer
-            implements WebApplicationInitializer {
-
-        @Override
-        public void onStartup(ServletContext servletContext)
-                throws ServletException {
-            AnnotationConfigWebApplicationContext context = new AnnotationConfigWebApplicationContext();
-            registerConfiguration(context);
-            servletContext.addListener(new ContextLoaderListener(context));
-
-            ServletRegistration.Dynamic registration = servletContext
-                    .addServlet("dispatcher", new SpringServlet(context));
-            registration.setLoadOnStartup(1);
-            registration.addMapping("/*");
-        }
-
-        private void registerConfiguration(
-                AnnotationConfigWebApplicationContext context) {
-            // register your configuration classes here
-        }
-    }
-
-    public class SampleWebAppInitializer extends VaadinMVCWebAppInitializer {
-
-        @Override
-        protected Collection<Class<?>> getConfigurationClasses() {
-            return Collections.singletonList(SampleConfiguration.class);
-        }
-    }
-
-    @Configuration
-    @ComponentScan
-    public class SampleConfiguration {
-    }
 }
